Magic 8 Ball version 2
I downloaded IntelliJ IDEA yesterday and spent quite a bit of time trying to work out how to open a new class and then get code to work. One of the tings I discovered was that I only needed to enter 'out.println();' to print.  I am not sure why 'System' is not required but as long as it works for now, I am happy.